synthetic_dataset_generator/requirements.txt
google-genai
flask>=2.0.1
python-dotenv>=0.19.0
pydantic>=2.0.0
watchdog>=2.1.5 